move LayeredReinforcedElements to Elements namespace
